[Bug fix] Param key type mismatch
In Ecto when you are attempting to cast params that have a mix of string
and atoms Ecto raises an error:
https://github.com/elixir-lang/ecto/blob/master/lib/ecto/changeset.ex#L428-L429

This fix will normalize all fields in the final param map to a string.
Phoenix params come in with string keys so this should be the preferred
normalization direction.